MS-419 Move index_file_size from IndexParam to TableSchema
Former-commit-id: 1626329551d8b8f9d25593c97006f2b789500f2b

+++ cpp/src/utils/ValidationUtil.cpp
@@ -76,7 +76,7 @@ ValidationUtil::ValidateTableIndexNlist(int32_t nlist) {
 }
 
 ServerError
-ValidationUtil::ValidateTableIndexFileSize(int32_t index_file_size) {
+ValidationUtil::ValidateTableIndexFileSize(int64_t index_file_size) {
     if(index_file_size <= 0 || index_file_size > index_file_size_limit) {
         return SERVER_INVALID_INDEX_FILE_SIZE;
     }
